requirements.txt
# AI Research Assistant - Configuration requise

# Python
python>=3.10

# LangGraph
langgraph>=0.0.20

# LangChain
langchain>=0.0.310
langchain-core>=0.1.4
langchain-groq>=0.1.0
langchain-text-splitters>=0.0.1

# ChromaDB
chromadb>=0.4.18

# API Clients
tavily-python>=0.2.6
serper-python>=0.1.3

# Utilitaires
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
markdown>=3.5.1
fpdf2>=2.7.5

# Streamlit (optionnel, pour l'interface utilisateur)
# streamlit>=1.28.0

# Tests
pytest>=7.0.0
pytest-asyncio>=0.21.0